requirements.txt
selenium==4.10.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pip install matplotlib
